Fix #25 (Gryphon crashed when renaming attributes of a data class)
When renaming attributes of a data class, references to this attribute
in the case start trigger have to be updated. This function assigned
the result of a `forEach` to the case start triggers, which unfortunatly
is `undefined`.
Also, updating references to changed attributes did not work previously.
This is now fixed!